requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.1.0
pydantic==2.10.6
